pyproject.toml
[project]
name = "protocol-omni"
version = "16.3.0"
description = "Protocol OMNI v16.3 - EAGLE EYE (Sentinel Audited 2026-01-24)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OMNI Team"}
]

dependencies = [
    "httpx>=0.28.0",
    "fastapi==0.128.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "langgraph>=1.0.3",
    "mem0ai>=1.0.2",
    "arize-phoenix-otel>=0.14.0",
    "opentelemetry-sdk>=1.39.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
